+++ src/main/java/com/sgi/pojos/Catbeneficiario.java
@@ -1,14 +1,11 @@
 package com.sgi.pojos;
 
 import java.io.Serializable;
-import javax.persistence.Basic;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
 import javax.persistence.Table;
 import javax.validation.constraints.Size;
 
@@ -21,7 +18,7 @@ public class Catbeneficiario implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)    
     @Column(name = "IdBen")
-    private Short idBen;
+    private Integer idBen;
     
     @Size(max = 40)
     @Column(name = "NomBen")
@@ -30,15 +27,15 @@ public class Catbeneficiario implements Serializable {
     public Catbeneficiario() {
     }
 
-    public Catbeneficiario(Short idBen) {
+    public Catbeneficiario(Integer idBen) {
         this.idBen = idBen;
     }
 
-    public Short getIdBen() {
+    public Integer getIdBen() {
         return idBen;
     }
 
-    public void setIdBen(Short idBen) {
+    public void setIdBen(Integer idBen) {
         this.idBen = idBen;
     }
