+++ src/main/java/com/sgi/pojos/Cdepto.java
@@ -1,21 +1,13 @@
 package com.sgi.pojos;
 
 import java.io.Serializable;
-import java.util.Collection;
-import javax.persistence.Basic;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
-import javax.persistence.OneToMany;
+import javax.persistence.Id;;
 import javax.persistence.OneToOne;
 import javax.persistence.PrimaryKeyJoinColumn;
 import javax.persistence.Table;
-import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
 @Entity
@@ -26,7 +18,7 @@ public class Cdepto implements Serializable {
     
     @Id    
     @Column(name = "idDpt")
-    private Short idDpt;
+    private Integer idDpt;
     
     @Size(max = 200)
     @Column(name = "NomDpt")
@@ -46,15 +38,15 @@ public class Cdepto implements Serializable {
     public Cdepto() {
     }
 
-    public Cdepto(Short idDpt) {
+    public Cdepto(Integer idDpt) {
         this.idDpt = idDpt;
     }
 
-    public Short getIdDpt() {
+    public Integer getIdDpt() {
         return idDpt;
     }
 
-    public void setIdDpt(Short idDpt) {
+    public void setIdDpt(Integer idDpt) {
         this.idDpt = idDpt;
     }
